Bugfix on account list, Javadoc, tagging as 0.1.3
A major bug in 0.1.2 is that accounts won't work properly (at all) if
there are network switches configured. This patch addresses that.

It also adds JavaDoc to each and every public method implemented.

Tagging as 0.1.3 release, the final 0.1.x release barring bugs and
moving HEAD to 0.2.x which will include drilldowns eventually.